Refactored bill and allocation
